Added labels to create a notice text fields. Added "Are you sure.." dialog boxes to delete swipes in Inbox and RequestsAdmin

+++ app/src/main/java/com/communityapp/inform/view/Inbox.java
@@ -6,6 +6,7 @@ import androidx.recyclerview.widget.ItemTouchHelper;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
+import android.app.AlertDialog;
 import android.app.ProgressDialog;
 import android.content.Intent;
 import android.os.Bundle;
@@ -72,8 +73,16 @@ public class Inbox extends AppCompatActivity {
 
             @Override
             public void onSwiped(@NonNull RecyclerView.ViewHolder viewHolder, int direction) {
-                adapter.deleteItem(viewHolder.getAdapterPosition());
-                Snackbar.make(relativeLayout, "Message deleted", Snackbar.LENGTH_LONG).setAction("Action", null).show();
+                new AlertDialog.Builder(Inbox.this)
+                        .setTitle("Delete Notification")
+                        .setMessage("Are you sure want to permanently delete this notification?")
+                        .setPositiveButton("Yes", (dialogInterface, i) -> {
+                            adapter.deleteItem(viewHolder.getAdapterPosition());
+                            Snackbar.make(relativeLayout, "Message deleted", Snackbar.LENGTH_LONG).setAction("Action", null).show();
+                        })
+                        .setNegativeButton("No", (dialogInterface, i) -> {
+                            //do nothing
+                        }).show();
             }
         }).attachToRecyclerView(msgRecyclerView);
